refactoring
- separate file for feedforward-layer and node
- delete unused function
